렌더링 최적화 및 성능 개선
RenderCommand 클래스를 추가하여 렌더링 명령을 효율적으로 관리합니다.
Core::App::Run()에서 렌더링 호출을 주석 처리하여 성능을 개선하고,
Dx11Main.cpp에 atomic 변수를 추가하여 멀티스레딩 동기화를 강화했습니다.
Camera.cpp의 PushRenderQueue() 함수 개선 및 새로운 SortRenderQueue() 함수 추가로
렌더링 큐 관리가 향상되었습니다.
LightMap.cpp에 코루틴 관련 함수 추가로 비동기 처리 성능을 개선하였으며,
SceneManager 및 SceneRenderer 클래스의 activeScene 접근 방식을 변경하여
멀티스레딩 안전성을 높였습니다.
Transform 클래스에 복사 및 이동 생성자 추가로 객체 관리가 개선되었고,
TimeSystem 클래스에서 프레임 카운트를 uint32에서 uint64로 변경하여
장기 실행 프로그램의 안정성을 높였습니다.

+++ Utility_Framework/TimeSystem.h
@@ -19,7 +19,7 @@ namespace DirectX11
 		double GetTotalSeconds() const { return TicksToSeconds(m_totalTicks); }
 
 		// Get total number of updates since start of the program.
-		uint32_t GetFrameCount() const { return m_frameCount; }
+		uint64_t GetFrameCount() const { return m_frameCount; }
 
 		// Get the current framerate.
 		uint32_t GetFramesPerSecond() const { return m_framesPerSecond; }
@@ -182,7 +182,7 @@ namespace DirectX11
 		uint32 m_fixedLeftOverTicks{};
 
 		// framerate 추적용 멤버입니다.
-		uint32 m_frameCount{};
+		uint64 m_frameCount{};
 		uint32 m_framesPerSecond{};
 		uint32 m_framesThisSecond{};
 		uint64 m_qpcSecondCounter{};
